chore: separate and reuse common tests
